requirements.txt
streamlit==1.52.1
openai==2.9.0
supabase==2.28.0
pandas==2.3.3
numpy==2.3.5
matplotlib==3.10.7
requests==2.32.5
python-dotenv==1.2.1
